chore: create explicit p2p versions of configuration if enabled
A bug that impacts node clients connecting with a particular behaviour, such as Ogmios, is present
in cardano-node@1.35.3, which at this point is the current production release. As a consequence,
p2p cannot be used, however, the configuration does enable p2p on non-production networks. This
change sets up an explicit set of configs for p2p, and transforms the defaults to use a static
topology, connecting to the IOG-managed relays directly, which is analogous to the _mainnet_ at
this point in time. To use p2p, append `_p2p` to the network name e.g. `preprod_p2p`